+++ libs/sdk-core/CHANGELOG.md
@@ -1,8 +1,10 @@
 # 1.5.0 (unreleased)
 
 ### Breaking changes
+
 - Remove `Search.SmartParagraph` interface as it's not used anymore
 - Rename `Search.SmartResult` interface to `Search.FieldResult` as `Smart` prefix doesn't make sense anymore
+- Add `selections` in field metadata
 
 # 1.4.0 (2023-06-23)